refactor: main에 mvp적용
