+++ app/src/main/java/woowacourse/movie/ui/main/MainActivity.kt
@@ -9,23 +9,20 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import woowacourse.movie.R
 import woowacourse.movie.domain.model.Movie
-import woowacourse.movie.domain.model.RunningTime
-import woowacourse.movie.domain.model.ScreeningDate
 import woowacourse.movie.ui.main.adapter.MoviesAdapter
 import woowacourse.movie.ui.reserve.ReserveActivity
 import woowacourse.movie.ui.reserve.ReserveActivity.Companion.KEY_RESERVE_ACTIVITY_MOVIE
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
 
-class MainActivity : AppCompatActivity() {
+class MainActivity : AppCompatActivity(), MainContract.View {
     private val moviesView: ListView by lazy { findViewById(R.id.lv_movies) }
+    private val presenter: MainContract.Presenter by lazy { MainPresenter(this) }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContentView(R.layout.activity_main)
         initSystemUI()
-        initMoviesView(movies())
+        presenter.showMovies()
     }
 
     private fun initSystemUI() {
@@ -36,20 +33,8 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    private fun movies(): List<Movie> {
-        val formatter: DateTimeFormatter =
-            DateTimeFormatter.ofPattern(getString(R.string.date_format))
-        val startDate: LocalDate = LocalDate.parse("2025.04.01", formatter)
-        val endDate: LocalDate = LocalDate.parse("2025.04.25", formatter)
-
-        return List(1000) {
-            Movie(
-                "해리포터와 마법사의 돌",
-                ScreeningDate(startDate, endDate),
-                RunningTime(152),
-                R.drawable.harrypotter,
-            )
-        }
+    override fun showMovies(movies: List<Movie>) {
+        initMoviesView(movies)
     }
 
     private fun initMoviesView(movies: List<Movie>) {
